refactor: create separate convertSchemaToColumns for easier generating UDT definitions
