+++ src/index.ts
@@ -27,6 +27,7 @@ export { driver };
 
 export { default as createAstraUri } from './createAstraUri';
 export { default as tableDefinitionFromSchema } from './tableDefinitionFromSchema';
+export { default as convertSchemaToColumns } from './convertSchemaToColumns';
 
 import * as AstraMongooseDriver from './driver';
 import type { Mongoose } from 'mongoose';
