+++ src/tableDefinitionFromSchema.ts
@@ -13,213 +13,15 @@
 // limitations under the License.
 
 import { CreateTableDefinition } from '@datastax/astra-db-ts';
-import { Schema, SchemaType } from 'mongoose';
-import { AstraMongooseError } from './astraMongooseError';
-
-type AllowedDataAPITypes = 'text' | 'double' | 'timestamp' | 'boolean' | 'decimal' | 'varint' | 'blob' | 'uuid' | 'int';
+import { Schema } from 'mongoose';
+import convertSchemaToColumns from './convertSchemaToColumns';
 
 /**
  * Given a Mongoose schema, create an equivalent Data API table definition for use with `createTable()`
  */
 export default function tableDefinitionFromSchema(schema: Schema): CreateTableDefinition {
-    const tableDefinition: CreateTableDefinition = {
+    return {
         primaryKey: '_id',
-        columns: {
-            _id: { type: 'text' },
-            __v: { type: 'int' }
-        }
+        columns: convertSchemaToColumns(schema)
     };
-    const schemaTypesForNestedPath: Record<string, SchemaType[]> = {};
-    for (const path of Object.keys(schema.paths)) {
-        const schemaType = schema.paths[path];
-
-        const type = mongooseTypeToDataAPIType(schemaType.instance);
-        const isNestedOrMap = path.indexOf('.') !== -1;
-        if (isNestedOrMap) {
-            const split = schemaType.path.split('.');
-            if (split.length > 2) {
-                throw new AstraMongooseError(`Cannot convert schema to Data API table definition: schemas with paths that are more than 2 levels deep are not supported (found ${path})`, {
-                    path,
-                    type,
-                    schema
-                });
-            }
-
-            // @ts-expect-error Mongoose TypeScript types don't have schema options
-            if (!schemaType.schema?.options?.udtName) {
-                // Maps of UDTs will be handled separately in the `instance === 'Map'` path
-                const nestedPath = split[0];
-                if (schemaTypesForNestedPath[nestedPath] == null) {
-                    schemaTypesForNestedPath[nestedPath] = [];
-                }
-                schemaTypesForNestedPath[nestedPath].push(schemaType);
-            }
-        } else if (type) {
-            tableDefinition.columns[path] = { type };
-        } else if (schemaType.instance === 'Array' || schemaType.instance === 'Vectorize') {
-            if (schemaType.schema) {
-                // @ts-expect-error Mongoose schemas don't have options property in TS
-                if (schemaType.schema.options.udtName) {
-                    tableDefinition.columns[path] = {
-                        type: 'list',
-                        // @ts-expect-error Astra-db-ts doesn't support UDT yet, and Mongoose schemas don't have options property in TS
-                        valueType: { type: 'userDefined', udtName: schemaType.schema.options.udtName }
-                    };
-                } else {
-                    throw new AstraMongooseError(`Cannot convert schema to Data API table definition: DocumentArray "${path}" is not supported`, {
-                        path,
-                        type,
-                        schema
-                    });
-                }
-
-            } else {
-                // Arrays always have an embedded schema type
-                const embeddedSchemaType = schemaType.getEmbeddedSchemaType() as SchemaType;
-                if (schemaType.options.dimension != null) {
-                    // If dimension, assume we're creating a vector column
-                    if (embeddedSchemaType.instance !== 'Number') {
-                        throw new AstraMongooseError(`Cannot convert schema to Data API table definition: vector column at "${path}" must be an array of numbers`, {
-                            path,
-                            type,
-                            schema
-                        });
-                    }
-                    tableDefinition.columns[path] = { type: 'vector', dimension: schemaType.options.dimension };
-                    if (schemaType.instance === 'Vectorize' && schemaType.options.service != null) {
-                        tableDefinition.columns[path].service = schemaType.options.service;
-                    }
-                } else {
-                    const valueType = mongooseTypeToDataAPIType(embeddedSchemaType.instance);
-                    if (valueType == null) {
-                        throw new AstraMongooseError(`Cannot convert schema to Data API table definition: unsupported array type at path "${path}"`, {
-                            path,
-                            valueType,
-                            type,
-                            schema
-                        });
-                    }
-                    tableDefinition.columns[path] = { type: 'list', valueType };
-                }
-            }
-        } else if (schemaType.instance === 'Embedded') {
-            // @ts-expect-error Mongoose schemas don't have options property in TS
-            if (schemaType.schema.options.udtName) {
-                // @ts-expect-error Astra-db-ts doesn't support UDT yet, and Mongoose schemas don't have options property in TS
-                tableDefinition.columns[path] = { type: 'userDefined', udtName: schemaType.schema.options.udtName };
-            } else {
-                const dataAPITypes: Set<AllowedDataAPITypes> = new Set();
-                for (const subpath of Object.keys(schemaType.schema.paths)) {
-                    const isNested = subpath.indexOf('.') !== -1;
-                    if (isNested) {
-                        throw new AstraMongooseError(`Cannot convert schema to Data API table definition: unsupported nested path underneath subdocument at path "${path}.${subpath}"`, {
-                            path,
-                            subpath,
-                            schema
-                        });
-                    }
-                    const type = mongooseTypeToDataAPIType(schemaType.schema.paths[subpath].instance);
-                    if (type == null) {
-                        throw new AstraMongooseError(`Cannot convert schema to Data API table definition: unsupported type in subdocument at path "${path}.${subpath}"`, {
-                            path,
-                            subpath,
-                            type,
-                            schema
-                        });
-                    }
-                    dataAPITypes.add(type);
-                }
-                // If all keys have same data type, then can just make map of that data type
-                if (dataAPITypes.size === 1) {
-                    tableDefinition.columns[path] = { type: 'map', keyType: 'text', valueType: [...dataAPITypes][0] };
-                } else {
-                    if (dataAPITypes.has('blob')) {
-                        throw new AstraMongooseError(`Cannot convert schema to Data API table definition: subdocuments with Buffer at "${path}" are not supported`, {
-                            path,
-                            type,
-                            schema
-                        });
-                    }
-                    tableDefinition.columns[path] = { type: 'map', keyType: 'text', valueType: 'text' };
-                }
-            }
-        } else if (schemaType.instance === 'Map') {
-            // @ts-expect-error Mongoose TypeScript types don't have schema options
-            if (schemaType.getEmbeddedSchemaType()?.schema?.options?.udtName) {
-                // Special handling for maps of UDTs
-                tableDefinition.columns[path] = {
-                    type: 'map',
-                    keyType: 'text',
-                    // @ts-expect-error Astra-db-ts doesn't support UDT yet
-                    valueType: {
-                        type: 'userDefined',
-                        // @ts-expect-error Mongoose TypeScript types don't have schema options
-                        udtName: schemaType.getEmbeddedSchemaType().schema.options.udtName
-                    }
-                };
-            }
-        } else {
-            throw new AstraMongooseError(`Cannot convert schema to Data API table definition: unsupported type at path "${path}"`, {
-                path,
-                schema
-            });
-        }
-    }
-
-    // Also handles maps
-    for (const nestedPath of Object.keys(schemaTypesForNestedPath)) {
-        const dataAPITypes: Set<AllowedDataAPITypes> = new Set();
-        for (const schemaType of schemaTypesForNestedPath[nestedPath]) {
-            const type = mongooseTypeToDataAPIType(schemaType.instance);
-            if (type == null) {
-                throw new AstraMongooseError(`Cannot convert schema to Data API table definition: unsupported type at path "${schemaType.path}"`, {
-                    nestedPath,
-                    type,
-                    schema
-                });
-            }
-            dataAPITypes.add(type);
-        }
-        if (dataAPITypes.has('blob')) {
-            throw new AstraMongooseError(`Cannot convert schema to Data API table definition: nested paths with Buffer at "${nestedPath}" are not supported`, {
-                nestedPath,
-                schema
-            });
-        }
-        // If all keys have same data type, then can just make map of that data type
-        if (dataAPITypes.size === 1) {
-            tableDefinition.columns[nestedPath] = { type: 'map', keyType: 'text', valueType: [...dataAPITypes][0] };
-        } else {
-            tableDefinition.columns[nestedPath] = { type: 'map', keyType: 'text', valueType: 'text' };
-        }
-    }
-
-    return tableDefinition;
-}
-
-function mongooseTypeToDataAPIType(type: string): AllowedDataAPITypes | null {
-    if (type === 'String') {
-        return 'text';
-    } else if (type === 'Number') {
-        return 'double';
-    } else if (type === 'Date') {
-        return 'timestamp';
-    } else if (type === 'Boolean') {
-        return 'boolean';
-    } else if (type === 'Decimal128') {
-        return 'decimal';
-    } else if (type === 'BigInt') {
-        return 'varint';
-    } else if (type === 'Buffer') {
-        return 'blob';
-    } else if (type === 'ObjectId') {
-        return 'text';
-    } else if (type === 'UUID') {
-        return 'uuid';
-    } else if (type === 'Int32') {
-        return 'int';
-    } else if (type === 'Double') {
-        return 'double';
-    }
-    return null;
 }
